AC .. after changing the data type to long long

+++ dsa-learning-series/lowsum.cpp
@@ -1,26 +1,26 @@
 #include<bits/stdc++.h>
 
-#define pr pair<int, int>
+#define pr pair<long long, long long>
 #define pb push_back
 #define mp make_pair
 
 using namespace std ;
 
 int main(){
-	int t , n , q, temp; 
+	long long t , n , q, temp; 
 	cin >> t ;
 	while (t--){
 		cin >> n >> q;
-		vector<int> a(n+1), b(n+1), ans(q+1); 
-		for (int i = 1 ; i <= n ; i++){
+		vector<long long> a(n+1), b(n+1), ans(q+1); 
+		for (long long i = 1 ; i <= n ; i++){
 			cin >> a[i] ;
 		}
-		for (int i = 1 ; i <= n ; i++){
+		for (long long i = 1 ; i <= n ; i++){
 			cin >> b[i] ;
 		}
 
-		vector<pair<int, int> > queries(q+1) ;
-		for(int i = 1 ; i <= q ; i++){
+		vector<pair<long long, long long> > queries(q+1) ;
+		for(long long i = 1 ; i <= q ; i++){
 			cin >> temp ;
 			queries[i] = make_pair(temp, i);
 		}
@@ -29,9 +29,9 @@ int main(){
 		sort(b.begin(), b.end());
 		sort(queries.begin(), queries.end());
 
-		vector<int> k(n+1);
+		vector<long long> k(n+1);
 		priority_queue<pr , vector<pr> , greater<pr> >  pq;
-		for(int i = 1 ; i <= n ;i ++){
+		for(long long i = 1 ; i <= n ;i ++){
 			k[i] = 1 ;
 			pq.push(mp(a[i]+b[k[i]], i) );
 		}
@@ -39,16 +39,16 @@ int main(){
 
 		pr curr = pq.top();
 		pq.pop() ;
-		int idx = curr.second ;
+		long long idx = curr.second ;
 		// cout << "value found : " << curr.first << endl ;
-		int cnt = 1 ;
+		long long cnt = 1 ;
 		k[idx]++ ;
 
 		if (k[idx] <= n){
 			pq.push(mp(a[idx] + b[k[idx]], idx));
 		}
 
-		for(int i = 1 ; i <= q ; i ++) {
+		for(long long i = 1 ; i <= q ; i ++) {
 			// cout << "here" ;
 			while(cnt < queries[i].first) {
 				curr = pq.top();
@@ -66,7 +66,7 @@ int main(){
 			ans[queries[i].second] = curr.first ;
 		}
 
-		for(int i = 1 ; i <= q ;i++){
+		for(long long i = 1 ; i <= q ;i++){
 			cout << ans[i] << endl ;
 		}
 	}
